fastapi-app/requirements.txt
fastapi
uvicorn
httpx
pytest
pytest-html
pytest-cov
prometheus-fastapi-instrumentator
prometheus-client
influxdb-client
python-logging-loki==0.3.1